requirements.txt
certifi==2017.11.5
chardet==3.0.4
idna==2.6
librouteros==1.0.4
python-pushover==0.3
requests==2.18.4
tinydb==3.7.0
ujson==1.35
urllib3==1.22